+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **1.0.1**
 * Fixed type _CompactIterable<String> is not a subtype of type List<String> for DefaultDonutModel
+* Support zoomKey, zoomType
 
 **1.0.0**
 * Dart 2 required